Refactor samplers namespaces
