+++ src/ImageProcessorCore/Samplers/Crop.cs
@@ -1,42 +1,68 @@
 ﻿// <copyright file="Crop.cs" company="James Jackson-South">
 // Copyright (c) James Jackson-South and contributors.
 // Licensed under the Apache License, Version 2.0.
-// </copyright>
+// </copyright>-------------------------------------------------------------------------------------------------------------------
 
-namespace ImageProcessorCore.Samplers
+namespace ImageProcessorCore
 {
-    using System.Threading.Tasks;
+    using Processors;
 
     /// <summary>
-    /// Provides methods to allow the cropping of an image.
+    /// Extension methods for the <see cref="Image"/> type.
     /// </summary>
-    public class Crop : ImageSampler
+    public static partial class ImageExtensions
     {
-        /// <inheritdoc/>
-        protected override void Apply(ImageBase target, ImageBase source, Rectangle targetRectangle, Rectangle sourceRectangle, int startY, int endY)
+        /// <summary>
+        /// Crops an image to the given width and height.
+        /// </summary>
+        /// <param name="source">The image to resize.</param>
+        /// <param name="width">The target image width.</param>
+        /// <param name="height">The target image height.</param>
+        /// <param name="progressHandler">A delegate which is called as progress is made processing the image.</param>
+        /// <returns>The <see cref="ImageProcessorCore.Image"/></returns>
+        public static Image Crop(this Image source, int width, int height, ProgressEventHandler progressHandler = null)
         {
-            int targetY = targetRectangle.Y;
-            int targetBottom = targetRectangle.Bottom;
-            int startX = targetRectangle.X;
-            int endX = targetRectangle.Right;
-            int sourceX = sourceRectangle.X;
-            int sourceY = sourceRectangle.Y;
+            return Crop(source, width, height, source.Bounds, progressHandler);
+        }
 
-            Parallel.For(
-            startY,
-            endY,
-            y =>
+        /// <summary>
+        /// Crops an image to the given width and height with the given source rectangle.
+        /// <remarks>
+        /// If the source rectangle is smaller than the target dimensions then the
+        /// area within the source is resized performing a zoomed crop.
+        /// </remarks>
+        /// </summary>
+        /// <param name="source">The image to crop.</param>
+        /// <param name="width">The target image width.</param>
+        /// <param name="height">The target image height.</param>
+        /// <param name="sourceRectangle">
+        /// The <see cref="Rectangle"/> structure that specifies the portion of the image object to draw.
+        /// </param>
+        /// <param name="progressHandler">A delegate which is called as progress is made processing the image.</param>
+        /// <returns>The <see cref="Image"/></returns>
+        public static Image Crop(this Image source, int width, int height, Rectangle sourceRectangle, ProgressEventHandler progressHandler = null)
+        {
+            Guard.MustBeGreaterThan(width, 0, nameof(width));
+            Guard.MustBeGreaterThan(height, 0, nameof(height));
+
+            if (sourceRectangle.Width < width || sourceRectangle.Height < height)
             {
-                if (y >= targetY && y < targetBottom)
-                {
-                    for (int x = startX; x < endX; x++)
-                    {
-                        target[x, y] = source[x + sourceX, y + sourceY];
-                    }
+                // If the source rectangle is smaller than the target perform a
+                // cropped zoom.
+                source = source.Resize(sourceRectangle.Width, sourceRectangle.Height);
+            }
 
-                    this.OnRowProcessed();
-                }
-            });
+            Crop processor = new Crop();
+            processor.OnProgress += progressHandler;
+
+            try
+            {
+                return source.Process(width, height, sourceRectangle, new Rectangle(0, 0, width, height), processor);
+            }
+            finally
+            {
+                processor.OnProgress -= progressHandler;
+            }
         }
     }
 }
