+++ src/ImageProcessorCore/Samplers/EntropyCrop.cs
@@ -1,102 +1,36 @@
 ﻿// <copyright file="EntropyCrop.cs" company="James Jackson-South">
 // Copyright (c) James Jackson-South and contributors.
 // Licensed under the Apache License, Version 2.0.
-// </copyright>
+// </copyright>-------------------------------------------------------------------------------------------------------------------
 
-namespace ImageProcessorCore.Samplers
+namespace ImageProcessorCore
 {
-    using System;
-    using System.Threading.Tasks;
-
-    using ImageProcessorCore.Filters;
+    using Processors;
 
     /// <summary>
-    /// Provides methods to allow the cropping of an image to preserve areas of highest
-    /// entropy.
+    /// Extension methods for the <see cref="Image"/> type.
     /// </summary>
-    public class EntropyCrop : ImageSampler
+    public static partial class ImageExtensions
     {
         /// <summary>
-        /// The rectangle for cropping
+        /// Crops an image to the area of greatest entropy.
         /// </summary>
-        private Rectangle cropRectangle;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="EntropyCrop"/> class.
-        /// </summary>
-        /// <param name="threshold">The threshold to split the image. Must be between 0 and 1.</param>
-        /// <exception cref="ArgumentException">
-        /// <paramref name="threshold"/> is less than 0 or is greater than 1.
-        /// </exception>
-        public EntropyCrop(float threshold)
+        /// <param name="source">The image to crop.</param>
+        /// <param name="threshold">The threshold for entropic density.</param>
+        /// <param name="progressHandler">A delegate which is called as progress is made processing the image.</param>
+        /// <returns>The <see cref="Image"/></returns>
+        public static Image EntropyCrop(this Image source, float threshold = .5f, ProgressEventHandler progressHandler = null)
         {
-            Guard.MustBeBetweenOrEqualTo(threshold, 0, 1, nameof(threshold));
-            this.Value = threshold;
-        }
+            EntropyCrop processor = new EntropyCrop(threshold);
+            processor.OnProgress += progressHandler;
 
-        /// <summary>
-        /// Gets the threshold value.
-        /// </summary>
-        public float Value { get; }
-
-        /// <inheritdoc/>
-        protected override void OnApply(ImageBase source, ImageBase target, Rectangle targetRectangle, Rectangle sourceRectangle)
-        {
-            using (ImageBase temp = new Image(source.Width, source.Height))
+            try
             {
-                // Detect the edges.
-                new Sobel().Apply(temp, source, sourceRectangle);
-
-                // Apply threshold binarization filter.
-                new Threshold(.5f).Apply(temp, temp, sourceRectangle);
-
-                // Search for the first white pixels
-                Rectangle rectangle = ImageMaths.GetFilteredBoundingRectangle(temp, 0);
-
-                // Reset the target pixel to the correct size.
-                target.SetPixels(rectangle.Width, rectangle.Height, new float[rectangle.Width * rectangle.Height * 4]);
-                this.cropRectangle = rectangle;
+                return source.Process(source.Width, source.Height, source.Bounds, Rectangle.Empty, processor);
             }
-        }
-
-        /// <inheritdoc/>
-        protected override void Apply(ImageBase target, ImageBase source, Rectangle targetRectangle, Rectangle sourceRectangle, int startY, int endY)
-        {
-            // Jump out, we'll deal with that later.
-            if (source.Bounds == target.Bounds)
+            finally
             {
-                return;
-            }
-
-            int targetY = this.cropRectangle.Y;
-            int targetBottom = this.cropRectangle.Bottom;
-            int startX = this.cropRectangle.X;
-            int endX = this.cropRectangle.Right;
-
-            Parallel.For(
-            startY,
-            endY,
-            y =>
-            {
-                if (y >= targetY && y < targetBottom)
-                {
-                    for (int x = startX; x < endX; x++)
-                    {
-                        target[x - startX, y - targetY] = source[x, y];
-                    }
-                }
-
-                this.OnRowProcessed();
-            });
-        }
-
-        /// <inheritdoc/>
-        protected override void AfterApply(ImageBase source, ImageBase target, Rectangle targetRectangle, Rectangle sourceRectangle)
-        {
-            // Copy the pixels over.
-            if (source.Bounds == target.Bounds)
-            {
-                target.ClonePixels(target.Width, target.Height, source.Pixels);
+                processor.OnProgress -= progressHandler;
             }
         }
     }
