+++ src/ImageProcessorCore/Samplers/Options/ResizeHelper.cs
@@ -3,7 +3,7 @@
 // Licensed under the Apache License, Version 2.0.
 // </copyright>
 
-namespace ImageProcessorCore.Samplers
+namespace ImageProcessorCore
 {
     using System;
     using System.Linq;
