+++ src/ImageProcessorCore/Samplers/Processors/ImageSampler.cs
@@ -3,7 +3,7 @@
 // Licensed under the Apache License, Version 2.0.
 // </copyright>
 
-namespace ImageProcessorCore.Samplers
+namespace ImageProcessorCore.Processors
 {
     /// <summary>
     /// Applies sampling methods to an image.
