+++ src/ImageProcessorCore/Samplers/Resamplers/HermiteResampler.cs
@@ -3,7 +3,7 @@
 // Licensed under the Apache License, Version 2.0.
 // </copyright>
 
-namespace ImageProcessorCore.Samplers
+namespace ImageProcessorCore.Processors
 {
     /// <summary>
     /// The function implements the hermite algorithm.
