+++ src/ImageProcessorCore/Samplers/Resamplers/Lanczos5Resampler.cs
@@ -3,7 +3,7 @@
 // Licensed under the Apache License, Version 2.0.
 // </copyright>
 
-namespace ImageProcessorCore.Samplers
+namespace ImageProcessorCore
 {
     /// <summary>
     /// The function implements the Lanczos kernel algorithm as described on
