+++ src/ImageProcessorCore/Samplers/Resize.cs
@@ -1,192 +1,134 @@
 ﻿// <copyright file="Resize.cs" company="James Jackson-South">
 // Copyright (c) James Jackson-South and contributors.
 // Licensed under the Apache License, Version 2.0.
-// </copyright>
+// </copyright>-------------------------------------------------------------------------------------------------------------------
 
-namespace ImageProcessorCore.Samplers
+namespace ImageProcessorCore
 {
-    using System.Threading.Tasks;
+    using Processors;
 
     /// <summary>
-    /// Provides methods that allow the resizing of images using various algorithms.
+    /// Extension methods for the <see cref="Image"/> type.
     /// </summary>
-    public class Resize : Resampler
+    public static partial class ImageExtensions
     {
         /// <summary>
-        /// The image used for storing the first pass pixels.
+        /// Resizes an image in accordance with the given <see cref="ResizeOptions"/>.
         /// </summary>
-        private Image firstPass;
+        /// <param name="source">The image to resize.</param>
+        /// <param name="options">The resize options.</param>
+        /// <param name="progressHandler">A delegate which is called as progress is made processing the image.</param>
+        /// <returns>The <see cref="Image"/></returns>
+        /// <remarks>Passing zero for one of height or width within the resize options will automatically preserve the aspect ratio of the original image</remarks>
+        public static Image Resize(this Image source, ResizeOptions options, ProgressEventHandler progressHandler = null)
+        {
+            // Ensure size is populated across both dimensions.
+            if (options.Size.Width == 0 && options.Size.Height > 0)
+            {
+                options.Size = new Size(source.Width * options.Size.Height / source.Height, options.Size.Height);
+            }
+
+            if (options.Size.Height == 0 && options.Size.Width > 0)
+            {
+                options.Size = new Size(options.Size.Width, source.Height * options.Size.Width / source.Width);
+            }
+
+            Rectangle targetRectangle = ResizeHelper.CalculateTargetLocationAndBounds(source, options);
+
+            return Resize(source, options.Size.Width, options.Size.Height, options.Sampler, source.Bounds, targetRectangle, options.Compand, progressHandler);
+        }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="Resize"/> class.
+        /// Resizes an image to the given width and height.
         /// </summary>
-        /// <param name="sampler">
-        /// The sampler to perform the resize operation.
+        /// <param name="source">The image to resize.</param>
+        /// <param name="width">The target image width.</param>
+        /// <param name="height">The target image height.</param>
+        /// <param name="progressHandler">A delegate which is called as progress is made processing the image.</param>
+        /// <returns>The <see cref="Image"/></returns>
+        /// <remarks>Passing zero for one of height or width will automatically preserve the aspect ratio of the original image</remarks>
+        public static Image Resize(this Image source, int width, int height, ProgressEventHandler progressHandler = null)
+        {
+            return Resize(source, width, height, new BicubicResampler(), false, progressHandler);
+        }
+
+        /// <summary>
+        /// Resizes an image to the given width and height.
+        /// </summary>
+        /// <param name="source">The image to resize.</param>
+        /// <param name="width">The target image width.</param>
+        /// <param name="height">The target image height.</param>
+        /// <param name="compand">Whether to compress and expand the image color-space to gamma correct the image during processing.</param>
+        /// <param name="progressHandler">A delegate which is called as progress is made processing the image.</param>
+        /// <returns>The <see cref="Image"/></returns>
+        /// <remarks>Passing zero for one of height or width will automatically preserve the aspect ratio of the original image</remarks>
+        public static Image Resize(this Image source, int width, int height, bool compand, ProgressEventHandler progressHandler = null)
+        {
+            return Resize(source, width, height, new BicubicResampler(), compand, progressHandler);
+        }
+
+        /// <summary>
+        /// Resizes an image to the given width and height with the given sampler.
+        /// </summary>
+        /// <param name="source">The image to resize.</param>
+        /// <param name="width">The target image width.</param>
+        /// <param name="height">The target image height.</param>
+        /// <param name="sampler">The <see cref="IResampler"/> to perform the resampling.</param>
+        /// <param name="compand">Whether to compress and expand the image color-space to gamma correct the image during processing.</param>
+        /// <param name="progressHandler">A delegate which is called as progress is made processing the image.</param>
+        /// <returns>The <see cref="Image"/></returns>
+        /// <remarks>Passing zero for one of height or width will automatically preserve the aspect ratio of the original image</remarks>
+        public static Image Resize(this Image source, int width, int height, IResampler sampler, bool compand, ProgressEventHandler progressHandler = null)
+        {
+            return Resize(source, width, height, sampler, source.Bounds, new Rectangle(0, 0, width, height), compand, progressHandler);
+        }
+
+        /// <summary>
+        /// Resizes an image to the given width and height with the given sampler and
+        /// source rectangle.
+        /// </summary>
+        /// <param name="source">The image to resize.</param>
+        /// <param name="width">The target image width.</param>
+        /// <param name="height">The target image height.</param>
+        /// <param name="sampler">The <see cref="IResampler"/> to perform the resampling.</param>
+        /// <param name="sourceRectangle">
+        /// The <see cref="Rectangle"/> structure that specifies the portion of the image object to draw.
         /// </param>
-        public Resize(IResampler sampler)
-            : base(sampler)
+        /// <param name="targetRectangle">
+        /// The <see cref="Rectangle"/> structure that specifies the portion of the target image object to draw to.
+        /// </param>
+        /// <param name="compand">Whether to compress and expand the image color-space to gamma correct the image during processing.</param>
+        /// <param name="progressHandler">A delegate which is called as progress is made processing the image.</param>
+        /// <returns>The <see cref="Image"/></returns>
+        /// <remarks>Passing zero for one of height or width will automatically preserve the aspect ratio of the original image</remarks>
+        public static Image Resize(this Image source, int width, int height, IResampler sampler, Rectangle sourceRectangle, Rectangle targetRectangle, bool compand = false, ProgressEventHandler progressHandler = null)
         {
-        }
-
-        /// <inheritdoc/>
-        public override int Parallelism { get; set; } = 1;
-
-        /// <inheritdoc/>
-        protected override void OnApply(ImageBase source, ImageBase target, Rectangle targetRectangle, Rectangle sourceRectangle)
-        {
-            if (!(this.Sampler is NearestNeighborResampler))
+            if (width == 0 && height > 0)
             {
-                this.HorizontalWeights = this.PrecomputeWeights(targetRectangle.Width, sourceRectangle.Width);
-                this.VerticalWeights = this.PrecomputeWeights(targetRectangle.Height, sourceRectangle.Height);
+                width = source.Width * height / source.Height;
+                targetRectangle.Width = width;
             }
 
-            this.firstPass = new Image(target.Width, source.Height);
-        }
-
-        /// <inheritdoc/>
-        protected override void Apply(ImageBase target, ImageBase source, Rectangle targetRectangle, Rectangle sourceRectangle, int startY, int endY)
-        {
-            // Jump out, we'll deal with that later.
-            if (source.Bounds == target.Bounds && sourceRectangle == targetRectangle)
+            if (height == 0 && width > 0)
             {
-                return;
+                height = source.Height * width / source.Width;
+                targetRectangle.Height = height;
             }
 
-            int width = target.Width;
-            int height = target.Height;
-            int sourceHeight = sourceRectangle.Height;
-            int targetX = target.Bounds.X;
-            int targetY = target.Bounds.Y;
-            int targetRight = target.Bounds.Right;
-            int targetBottom = target.Bounds.Bottom;
-            int startX = targetRectangle.X;
-            int endX = targetRectangle.Right;
-            bool compand = this.Compand;
+            Guard.MustBeGreaterThan(width, 0, nameof(width));
+            Guard.MustBeGreaterThan(height, 0, nameof(height));
 
-            if (this.Sampler is NearestNeighborResampler)
+            Resize processor = new Resize(sampler) { Compand = compand };
+            processor.OnProgress += progressHandler;
+
+            try
             {
-                // Scaling factors
-                float widthFactor = sourceRectangle.Width / (float)targetRectangle.Width;
-                float heightFactor = sourceRectangle.Height / (float)targetRectangle.Height;
-
-                Parallel.For(
-                    startY,
-                    endY,
-                    y =>
-                    {
-                        if (targetY <= y && y < targetBottom)
-                        {
-                            // Y coordinates of source points
-                            int originY = (int)((y - startY) * heightFactor);
-
-                            for (int x = startX; x < endX; x++)
-                            {
-                                if (targetX <= x && x < targetRight)
-                                {
-                                    // X coordinates of source points
-                                    int originX = (int)((x - startX) * widthFactor);
-
-                                    target[x, y] = source[originX, originY];
-                                }
-                            }
-
-                            this.OnRowProcessed();
-                        }
-                    });
-
-                // Break out now.
-                return;
+                return source.Process(width, height, sourceRectangle, targetRectangle, processor);
             }
-
-            // Interpolate the image using the calculated weights.
-            // A 2-pass 1D algorithm appears to be faster than splitting a 1-pass 2D algorithm 
-            // First process the columns. Since we are not using multiple threads startY and endY
-            // are the upper and lower bounds of the source rectangle.
-            Parallel.For(
-                0,
-                sourceHeight,
-                y =>
-                    {
-                        for (int x = startX; x < endX; x++)
-                        {
-                            if (x >= 0 && x < width)
-                            {
-                                // Ensure offsets are normalised for cropping and padding.
-                                int offsetX = x - startX;
-                                float sum = this.HorizontalWeights[offsetX].Sum;
-                                Weight[] horizontalValues = this.HorizontalWeights[offsetX].Values;
-
-                                // Destination color components
-                                Color destination = new Color();
-
-                                for (int i = 0; i < sum; i++)
-                                {
-                                    Weight xw = horizontalValues[i];
-                                    int originX = xw.Index;
-                                    Color sourceColor = compand ? Color.Expand(source[originX, y]) : source[originX, y];
-                                    destination += sourceColor * xw.Value;
-                                }
-
-                                if (compand)
-                                {
-                                    destination = Color.Compress(destination);
-                                }
-
-                                this.firstPass[x, y] = destination;
-                            }
-                        }
-                    });
-
-            // Now process the rows.
-            Parallel.For(
-                startY,
-                endY,
-                y =>
-                    {
-                        if (y >= 0 && y < height)
-                        {
-                            // Ensure offsets are normalised for cropping and padding.
-                            int offsetY = y - startY;
-                            float sum = this.VerticalWeights[offsetY].Sum;
-                            Weight[] verticalValues = this.VerticalWeights[offsetY].Values;
-
-                            for (int x = 0; x < width; x++)
-                            {
-                                // Destination color components
-                                Color destination = new Color();
-
-                                for (int i = 0; i < sum; i++)
-                                {
-                                    Weight yw = verticalValues[i];
-                                    int originY = yw.Index;
-                                    Color sourceColor = compand ? Color.Expand(this.firstPass[x, originY]) : this.firstPass[x, originY];
-                                    destination += sourceColor * yw.Value;
-                                }
-
-                                if (compand)
-                                {
-                                    destination = Color.Compress(destination);
-                                }
-
-                                target[x, y] = destination;
-                            }
-                        }
-
-                        this.OnRowProcessed();
-                    });
-        }
-
-        /// <inheritdoc/>
-        protected override void AfterApply(ImageBase source, ImageBase target, Rectangle targetRectangle, Rectangle sourceRectangle)
-        {
-            // Copy the pixels over.
-            if (source.Bounds == target.Bounds && sourceRectangle == targetRectangle)
+            finally
             {
-                target.ClonePixels(target.Width, target.Height, source.Pixels);
+                processor.OnProgress -= progressHandler;
             }
-
-            // Clean up
-            this.firstPass?.Dispose();
         }
     }
 }
